+++ src/CommandingOfficers/DefendPeace/misc/Bear_Bull.java
@@ -1,8 +1,13 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.misc;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderAbility.CostBasis;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameScenario;
 import Engine.XYCoord;
 import Engine.Combat.DamagePopup;
