+++ src/CommandingOfficers/DefendPeace/misc/DocLight.java
@@ -1,6 +1,10 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.misc;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.DeployableCommander;
 import CommandingOfficers.AW4.RuinedCommander;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameScenario;
 import Engine.Combat.StrikeParams.BattleParams;
 import Engine.GameEvents.GameEventQueue;
