+++ src/CommandingOfficers/DefendPeace/misc/Venge.java
@@ -1,7 +1,12 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.misc;
 
 import java.util.ArrayList;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderAbility.CostBasis;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.Army;
 import Engine.GameScenario;
 import Engine.Combat.StrikeParams;
